refactor: 쿼리키 선언 및 폴더구조 수정
