+++ maeil-mail/src/hooks/queries/keys/queryKeys.ts
@@ -1,0 +1,7 @@
+import { Category } from '@/types';
+
+export const questionKeys = {
+  all: ['questions'],
+  detail: (id: string) => [questionKeys.all, id],
+  category: (category: Category) => [questionKeys.all, category],
+};
